refactor: Move hot compaction function into hot compaction module
